+++ examples/resources/iosxe_vrrp_ipv6/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxe_vrrp_ipv6.example "<type>,<name>,<group_id>"