+++ internal/provider/resource_iosxe_vrrp_ipv6_test.go
@@ -33,40 +33,39 @@ import (
 
 // Section below is generated&owned by "gen/generator.go". //template:begin testAcc
 
-func TestAccIosxeVRRP(t *testing.T) {
+func TestAccIosxeVRRPIPv6(t *testing.T) {
 	if os.Getenv("C8000V") == "" {
 		t.Skip("skipping test, set environment variable C8000V")
 	}
 	var checks []resource.TestCheckFunc
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp.test", "group_id", "1"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp.test", "address_primary_address", "192.0.2.254"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp.test", "address_primary", "true"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp.test", "secondary_addresses.0.address", "192.0.2.253"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp.test", "secondary_addresses.0.secondary", "true"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp.test", "priority", "110"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp.test", "preempt_delay_minimum", "30"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp.test", "timers_advertise", "3000"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp.test", "description", "VRRP-GROUP-1"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp.test", "tracks.0.object_id", "1"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp.test", "tracks.0.decrement", "20"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp.test", "tracks.0.shutdown", "false"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp.test", "shutdown", "false"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp_ipv6.test", "group_id", "2"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp_ipv6.test", "ipv6_link_local", "FE80::1"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp_ipv6.test", "ipv6_primary", "true"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp_ipv6.test", "ipv6_prefixes.0.prefix", "2001:DB8::FFFF/64"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp_ipv6.test", "priority", "110"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp_ipv6.test", "preempt_delay_minimum", "30"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp_ipv6.test", "timers_advertise", "3000"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp_ipv6.test", "description", "VRRP-IPV6-GROUP-2"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp_ipv6.test", "tracks.0.object_id", "1"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp_ipv6.test", "tracks.0.decrement", "20"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp_ipv6.test", "tracks.0.shutdown", "false"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_vrrp_ipv6.test", "shutdown", "false"))
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccIosxeVRRPPrerequisitesConfig + testAccIosxeVRRPConfig_minimum(),
+				Config: testAccIosxeVRRPIPv6PrerequisitesConfig + testAccIosxeVRRPIPv6Config_minimum(),
 			},
 			{
-				Config: testAccIosxeVRRPPrerequisitesConfig + testAccIosxeVRRPConfig_all(),
+				Config: testAccIosxeVRRPIPv6PrerequisitesConfig + testAccIosxeVRRPIPv6Config_all(),
 				Check:  resource.ComposeTestCheckFunc(checks...),
 			},
 			{
-				ResourceName:            "iosxe_vrrp.test",
+				ResourceName:            "iosxe_vrrp_ipv6.test",
 				ImportState:             true,
 				ImportStateVerify:       true,
-				ImportStateIdFunc:       iosxeVRRPImportStateIdFunc("iosxe_vrrp.test"),
+				ImportStateIdFunc:       iosxeVRRPIPv6ImportStateIdFunc("iosxe_vrrp_ipv6.test"),
 				ImportStateVerifyIgnore: []string{},
 				Check:                   resource.ComposeTestCheckFunc(checks...),
 			},
@@ -78,7 +77,7 @@ func TestAccIosxeVRRP(t *testing.T) {
 
 // Section below is generated&owned by "gen/generator.go". //template:begin importStateIdFunc
 
-func iosxeVRRPImportStateIdFunc(resourceName string) resource.ImportStateIdFunc {
+func iosxeVRRPIPv6ImportStateIdFunc(resourceName string) resource.ImportStateIdFunc {
 	return func(s *terraform.State) (string, error) {
 		primary := s.RootModule().Resources[resourceName].Primary
 		Type := primary.Attributes["type"]
@@ -92,7 +91,7 @@ func iosxeVRRPImportStateIdFunc(resourceName string) resource.ImportStateIdFunc 
 // End of section. //template:end importStateIdFunc
 
 // Section below is generated&owned by "gen/generator.go". //template:begin testPrerequisites
-const testAccIosxeVRRPPrerequisitesConfig = `
+const testAccIosxeVRRPIPv6PrerequisitesConfig = `
 resource "iosxe_yang" "PreReq0" {
 	path = "/Cisco-IOS-XE-native:native/fhrp/version"
 	attributes = {
@@ -108,10 +107,9 @@ resource "iosxe_yang" "PreReq1" {
 }
 
 resource "iosxe_yang" "PreReq2" {
-	path = "/Cisco-IOS-XE-native:native/interface/Vlan[name=123]/ip/address/primary"
+	path = "/Cisco-IOS-XE-native:native/interface/Vlan[name=123]/ipv6/Cisco-IOS-XE-ipv6:address/link-local-address[address=FE80::1]"
 	attributes = {
-		"address" = "192.0.2.1"
-		"mask" = "255.255.255.0"
+		"address" = "FE80::1"
 	}
 }
 
@@ -121,12 +119,12 @@ resource "iosxe_yang" "PreReq2" {
 
 // Section below is generated&owned by "gen/generator.go". //template:begin testAccConfigMinimal
 
-func testAccIosxeVRRPConfig_minimum() string {
-	config := `resource "iosxe_vrrp" "test" {` + "\n"
+func testAccIosxeVRRPIPv6Config_minimum() string {
+	config := `resource "iosxe_vrrp_ipv6" "test" {` + "\n"
 	config += `	type = "Vlan"` + "\n"
 	config += `	name = "123"` + "\n"
-	config += `	group_id = 1` + "\n"
-	config += `	address_primary_address = "192.0.2.254"` + "\n"
+	config += `	group_id = 2` + "\n"
+	config += `	ipv6_link_local = "FE80::1"` + "\n"
 	config += `	depends_on = [iosxe_yang.PreReq0, iosxe_yang.PreReq1, iosxe_yang.PreReq2, ]` + "\n"
 	config += `}` + "\n"
 	return config
@@ -136,21 +134,20 @@ func testAccIosxeVRRPConfig_minimum() string {
 
 // Section below is generated&owned by "gen/generator.go". //template:begin testAccConfigAll
 
-func testAccIosxeVRRPConfig_all() string {
-	config := `resource "iosxe_vrrp" "test" {` + "\n"
+func testAccIosxeVRRPIPv6Config_all() string {
+	config := `resource "iosxe_vrrp_ipv6" "test" {` + "\n"
 	config += `	type = "Vlan"` + "\n"
 	config += `	name = "123"` + "\n"
-	config += `	group_id = 1` + "\n"
-	config += `	address_primary_address = "192.0.2.254"` + "\n"
-	config += `	address_primary = true` + "\n"
-	config += `	secondary_addresses = [{` + "\n"
-	config += `		address = "192.0.2.253"` + "\n"
-	config += `		secondary = true` + "\n"
+	config += `	group_id = 2` + "\n"
+	config += `	ipv6_link_local = "FE80::1"` + "\n"
+	config += `	ipv6_primary = true` + "\n"
+	config += `	ipv6_prefixes = [{` + "\n"
+	config += `		prefix = "2001:DB8::FFFF/64"` + "\n"
 	config += `	}]` + "\n"
 	config += `	priority = 110` + "\n"
 	config += `	preempt_delay_minimum = 30` + "\n"
 	config += `	timers_advertise = 3000` + "\n"
-	config += `	description = "VRRP-GROUP-1"` + "\n"
+	config += `	description = "VRRP-IPV6-GROUP-2"` + "\n"
 	config += `	tracks = [{` + "\n"
 	config += `		object_id = "1"` + "\n"
 	config += `		decrement = 20` + "\n"